@@ -9,8 +9,8 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "emailjs-com": "^3.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-google-recaptcha": "^2.1.0",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
